[RFC] Introduce Socket32 class
After @dhilanr and I took a look at the code for pset 2, we felt that it would likely be pretty intimidating for students.

Here's a pass I'm taking at simplifying pset2 - instead using a `Socket32` class which abstracts away a lot of the flags, etc. present in the default socket class.

I've also added in some basic typechecking and pretty-printing functionality to give helpful error messages when students make mistakes on using the `Socket32` class.